- Histogram plugin's test histogram (shows tracks in an event) now uses the same widget for displaying min/max/bins as the combination histograms. - Histogram plugin does not load the combo histogram if AGeometry is not present - Possible to setup an "optional" dependency requirement for plugins. Used by histogram plugin

+++ src/ahist/AComboHistogram.cpp
@@ -28,7 +28,7 @@ AComboHistogram::AComboHistogram(QString _name,QWidget* parent):QWidget(parent),
 	  this,SLOT(updateBins()));
 }
 
-void AComboHistogram::addData(int point)
+void AComboHistogram::addData(float point)
 {
   data.push_back(point);
   hist->Fill(point);
